model-deployment/requirements.txt
gradio==5.33.2
Pillow==11.2.1
torch==2.7.0
torchvision==0.22.0
